Дополнена логика создания команд

+++ src/main/java/ru/hits/attackdefenceplatform/core/team/repository/TeamEntity.java
@@ -19,5 +19,5 @@ public class TeamEntity {
 
     private String name;
 
-    private Integer maxMembers;
+    private Long maxMembers;
 }
